refactor: refactored some of the code
