+++ packages/amplify-category-function/src/provider-utils/awscloudformation/service-walkthroughs/lambda-walkthrough.ts
@@ -156,7 +156,6 @@ export async function updateWalkthrough(context, lambdaToUpdate?: string) {
       currentParameters,
       functionParameters.environmentMap,
       functionParameters.mutableParametersState,
-      context.amplify.getProjectMeta(),
     );
 
     context.amplify.writeObjectAsJson(cfnFilePath, cfnContent, true);
